requirements.txt
openai
google-genai
matplotlib
numpy
pandas